+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Detailed description for NumConns (CASSGO-3)
 
+- Change Batch API to be consistent with Query() (CASSGO-7)
+
 ### Fixed
 
 - Retry policy now takes into account query idempotency (CASSGO-27)